requirements.txt
asgiref==3.11.0
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
defusedxml==0.7.1
Django==6.0.1
django-filter==25.2
djangorestframework==3.16.1
drf-yasg==1.21.14
httpie==3.2.4
idna==3.11
importlib_resources==6.5.2
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.0
packaging==26.0
pyaml==25.7.0
Pygments==2.19.2
PySocks==1.7.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.30.0
setuptools==80.9.0
sqlparse==0.5.5
swagger-spec-validator==3.0.4
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
